added cloudiinary for uploading pizza pics and applied token on creating and updatind and deleting method

+++ pages/api/products/[id].js
@@ -5,7 +5,10 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
+
+  const token = cookies.token;
 
   await dbConnect();
 
@@ -21,6 +24,9 @@ export default async function handler(req, res) {
 
   //*Post method
   if (method === "PUT") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authenticated");
+    }
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
@@ -33,6 +39,9 @@ export default async function handler(req, res) {
 
   //*DELETE method
   if (method === "DELETE") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authenticated");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(201).json("The product has been deleted!");
